Framework seperation
