+++ tests/test_main.py
@@ -3,22 +3,23 @@ import tempfile
 
 import pytest
 
-from apistar import App, exceptions
+from apistar import exceptions
+from apistar.interfaces import App
 from apistar.main import default_app, load_app
-
-
-def test_load_app():
-    with tempfile.TemporaryDirectory() as tempdir:
-        os.chdir(tempdir)
-        app = App()
-        app.main(['new', '.', '--layout', 'minimal'], standalone_mode=False)
-        loaded = load_app()
-        assert isinstance(loaded, App)
 
 
 def test_default_app():
     default = default_app()
     assert isinstance(default, App)
+
+
+def test_load_app():
+    with tempfile.TemporaryDirectory() as tempdir:
+        os.chdir(tempdir)
+        app = default_app()
+        app.main(['new', '.', '--layout', 'minimal'], standalone_mode=False)
+        loaded = load_app()
+        assert isinstance(loaded, App)
 
 
 def test_load_misconfigured_app():
